+++ lib/cryptozoologist/dictionaries.rb
@@ -27,6 +27,10 @@ module Cryptozoologist
         colors: {
           paint: Colors::Paint,
           web: Colors::WebSafe
+        },
+        cities: {
+          words: Cities::Words,
+          terminologies: Cities::Terminologies
         }
       }
     end
@@ -45,6 +49,9 @@ module Cryptozoologist
 
     def states
       States.list
+
+    def cities
+      create_list(:cities)
     end
 
     private
@@ -54,8 +61,8 @@ module Cryptozoologist
       list.flatten
     end
 
-    def filter_library(key)
-      dictionaries = library[key].reject do |key, value|
+    def filter_library(filter_key)
+      dictionaries = library[filter_key].reject do |key, _value|
         Cryptozoologist.configuration.exclude.include?(key)
       end
       dictionaries.values