+++ lib/cryptozoologist/dictionaries/cities/teminologies.rb
@@ -1,0 +1,9 @@
+module Cryptozoologist
+  module Cities
+    module Terminologies
+      def self.list
+        ['burg', 'field', 'land', 'landia', 'port', 'side', 'town', 'ville']
+      end
+    end
+  end
+end